requirements.txt
termcolor>=1.1.0
py3dns>=2.3.6
